Simplify Android video core naming
