+++ mobile/packages/native_video_editor/android/src/main/kotlin/io/ente/native_video_editor/core/VideoExportEngine.kt
@@ -35,9 +35,9 @@ import kotlinx.coroutines.withContext
 import java.io.File
 
 @UnstableApi
-class AndroidVideoExportEngine(
+class VideoExportEngine(
     private val context: Context,
-    private val metadataReader: AndroidVideoMetadataReader
+    private val metadataReader: VideoMetadataReader
 ) {
     suspend fun processVideo(
         request: VideoEditRequest,
